+++ src/it/scala/stellar/sdk/LocalNetworkIntegrationSpec.scala
@@ -618,6 +618,7 @@ class LocalNetworkIntegrationSpec(implicit ee: ExecutionEnv) extends Specificati
       AbsolutelyBefore(instant),
       Unconditional
     )
+    val id = ClaimableBalanceHashId(ByteString.decodeHex("a20bec491b3901338a39b1430c4ca177176641698a267c1c92df5aaf8b9688ee"))
 
     "be creatable" >> {
       for {
@@ -644,9 +645,11 @@ class LocalNetworkIntegrationSpec(implicit ee: ExecutionEnv) extends Specificati
       }
     }
 
-    def includeTheOneWeJustPosted = {
-      beLike[List[ClaimableBalance]] { case List(only) =>
-        only.id mustEqual ByteString.decodeHex("a20bec491b3901338a39b1430c4ca177176641698a267c1c92df5aaf8b9688ee")
+    def includeTheOneWeJustPosted =
+      beLike[Seq[ClaimableBalance]] { case seq =>
+        seq must haveSize(1)
+        val only = seq.head
+        only.id mustEqual id
         only.amount mustEqual Amount(5000, dachshundB)
         only.sponsor mustEqual accnB
         only.claimants must containTheSameElementsAs(List(
@@ -654,33 +657,30 @@ class LocalNetworkIntegrationSpec(implicit ee: ExecutionEnv) extends Specificati
           AccountIdClaimant(accnC.asPublicKey, Unconditional)
         ))
       }
-    }
 
     "be listable by account" >> {
-      for {
-        balances <- network.claimsByClaimant(accnA)
-      } yield balances.toList must includeTheOneWeJustPosted
+      network.claimsByClaimant(accnA) must includeTheOneWeJustPosted.awaitFor(10.seconds)
     }
 
     "be listable by asset" >> {
-      for {
-        balances <- network.claimsByAsset(dachshundB)
-      } yield balances.toList must includeTheOneWeJustPosted
+      network.claimsByAsset(dachshundB) must includeTheOneWeJustPosted.awaitFor(10.seconds)
     }
 
     "be listable by sponsor" >> {
-      for {
-        balances <- network.claimsBySponsor(accnB)
-      } yield balances.toList must includeTheOneWeJustPosted
+      network.claimsBySponsor(accnB) must includeTheOneWeJustPosted.awaitFor(10.seconds)
+    }
+
+    "be findable by id" >> {
+      Await.result(network.claim(id).map(LazyList(_)), 10.seconds) must includeTheOneWeJustPosted
     }
 
     "be claimable" >> {
       for {
         account <- network.account(accnA)
-        hash <- network.claimsByClaimant(accnA).map(_.head.id)
+        id <- network.claimsByClaimant(accnA).map(_.head.id)
         txn = Transaction(
           source = account,
-          operations = List(ClaimClaimableBalanceOperation(ClaimableBalanceHashId(hash))),
+          operations = List(ClaimClaimableBalanceOperation(id)),
           timeBounds = TimeBounds.Unbounded,
           maxFee = NativeAmount(100)
         ).sign(accnA)